refactor: modularize components
